https://github.com/favoyang/unity-addressable-importer/pull/33 fixes
