--- Runtime/Unity.AddressableImporter.Runtime.asmdef
@@ -1,3 +1,0 @@
-﻿{
-	"name": "Unity.AddressableImporter.Runtime"
-}
