--- Runtime/Unity.AddressableImporter.Runtime.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 99e99866ebf58a5438e15861a180c74f
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
